+++ src/routers/commentsRouter.js
@@ -1,9 +1,10 @@
 import express from 'express';
-import { getComments } from '../controllers/getCommentsController.js';
+import { insertPostComment, getComments } from '../controllers/commentsController.js';
 import { tokenVerification } from '../middlewares/tokenValidationMiddleware.js';
 
 const router = express.Router()
 
 router.get('/comments/:postId', tokenVerification, getComments)
+router.post('/comments/:postId', tokenVerification, insertPostComment)
 
 export default router;
